+++ core/use_cases/diets/calculate_diet_amount.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from core.repositories.diet_service_repository import DietServiceRepository
 
 class CalculateDietAmountUseCase:
@@ -15,22 +14,22 @@ class CalculateDietAmountUseCase:
         dinner_count: int,
         accommodation_count: int,
         accommodation_payment_method: str
-    ) -> Decimal:
+    ) -> float:
         # Obtener los precios según localidad
         diet_service = self.diet_service_repository.get_by_local(is_local)
         if not diet_service:
             raise ValueError("No se encontraron precios para la localidad especificada")
         
         # Calcular montos
-        breakfast_total = Decimal(diet_service.breakfast_price) * breakfast_count
-        lunch_total = Decimal(diet_service.lunch_price) * lunch_count
-        dinner_total = Decimal(diet_service.dinner_price) * dinner_count
+        breakfast_total = float(diet_service.breakfast_price) * breakfast_count
+        lunch_total = float(diet_service.lunch_price) * lunch_count
+        dinner_total = float(diet_service.dinner_price) * dinner_count
         
         # Calcular alojamiento según método de pago
         if accommodation_payment_method == "CASH":
-            accommodation_total = Decimal(diet_service.accommodation_cash_price) * accommodation_count
+            accommodation_total = float(diet_service.accommodation_cash_price) * accommodation_count
         else:
-            accommodation_total = Decimal(diet_service.accommodation_card_price) * accommodation_count
+            accommodation_total = float(diet_service.accommodation_card_price) * accommodation_count
         
         total_amount = breakfast_total + lunch_total + dinner_total + accommodation_total
